feat(header): implement conditional header rendering based on route
Add logic to the Header component to conditionally render based on the route's hideHeader property. Update tests to verify header visibility for different routes. Remove unused layout tests.
