+++ front/src/components/Layout/Header/Header.tsx
@@ -3,14 +3,27 @@ import AddCalendarIcon from '@material-symbols/svg-400/outlined/calendar_add_on.
 import Person from '@material-symbols/svg-400/outlined/person.svg?react';
 import logo from '../../../../public/assets/logo.png';
 
+import { RouteHandle } from '@Front/routing/routeHandle';
+import { useMemo } from 'react';
+import { UIMatch, useMatches } from 'react-router';
 import './Header.scss';
 
-export const Header = () => (
-  <header className="header">
-    <img src={logo} alt="Slot Finder logo" className="header__logo" />
-    <div className="header__buttons">
-      <Button icon={AddCalendarIcon} variant="secondary" aria-label="add event" className="header__button" />
-      <Button icon={Person} variant="secondary" aria-label="user profile" className="header__button" />
-    </div>
-  </header>
-);
+export const Header = () => {
+  const matches = useMatches() as UIMatch<unknown, RouteHandle>[];
+
+  const hideHeader = useMemo(() => matches.some(match => match.handle?.hideHeader === true), [matches]);
+
+  if (hideHeader) {
+    return null;
+  }
+
+  return (
+    <header className="header">
+      <img src={logo} alt="Slot Finder logo" className="header__logo" />
+      <div className="header__buttons">
+        <Button icon={AddCalendarIcon} variant="secondary" aria-label="add event" className="header__button" />
+        <Button icon={Person} variant="secondary" aria-label="user profile" className="header__button" />
+      </div>
+    </header>
+  );
+};
